requirements.txt
pygame
numpy
pytorch